Rename BalancesService to BalancesUserService in module and controller files
