+++ src/resources/financial/balances/balances.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { PrismaModule } from '../../prisma/prisma.module';
 import { UsersModule } from 'src/resources/users/users.module';
-import { BalancesService } from './balances.service';
+import { BalancesService } from './balances-user.service';
 import { BalancesUserController } from './balances.user.controller';
 import { BalancesAdminController } from './balances.admin.controller';
 import { BalancesAdminService } from './balances-admin.service';
